// src/app/models/FeederModels.ts
import { MRID } from './MRID';

export interface FeederModels {
  regions: Array<{ regionName: string; regionID: string; index: number }>;
  subregions: Array<{ subregionName: string; subregionID: string; index: number }>;
  lines: Array<{ name: string; mRID: string; index: number }>;
  mRIDs: MRID[];
}